requirements.txt
dj-database-url==0.4.2
Django==2.2.28
django-crispy-forms==1.13.0
gunicorn==19.7.1
psycopg2==2.7.4
pytz==2018.3
sqlparse==0.4.2
whitenoise==3.3.1
